face_cut/requirements.txt
boto3==1.13.15
ydb==3.5.2
Pillow==9.5.0
